requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
attrs==25.3.0
    # via
    #   cattrs
    #   jsii
    #   jsonschema
    #   referencing
aws-cdk-asset-awscli-v1==2.2.231
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk-cloud-assembly-schema==41.2.0
    # via aws-cdk-lib
aws-cdk-lib==2.189.0
    # via bedrock-knowledge-bot (pyproject.toml)
beautifulsoup4==4.13.3
    # via favicon
blinker==1.9.0
    # via streamlit
boto3==1.37.33
    # via bedrock-knowledge-bot (pyproject.toml)
botocore==1.37.33
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via streamlit
cattrs==24.1.3
    # via jsii
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
constructs==10.4.2
    # via aws-cdk-lib
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
entrypoints==0.4
    # via streamlit-extras
faker==37.1.0
    # via streamlit-faker
favicon==0.7.0
    # via markdownlit
fonttools==4.57.0
    # via matplotlib
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
htbuilder==0.9.0
    # via
    #   markdownlit
    #   st-annotated-text
    #   streamlit-extras
idna==3.10
    # via requests
importlib-resources==6.5.2
    # via jsii
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   streamlit-camera-input-live
    #   streamlit-image-coordinates
    #   streamlit-keyup
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.111.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.8
    # via matplotlib
lxml==5.3.2
    # via markdownlit
markdown==3.8
    # via
    #   markdownlit
    #   pymdown-extensions
markdownlit==0.0.7
    # via streamlit-extras
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.1
    # via streamlit-faker
narwhals==1.34.1
    # via
    #   altair
    #   plotly
numpy==2.2.4
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   matplotlib
    #   plotly
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.2.1
    # via
    #   matplotlib
    #   streamlit
plotly==6.0.1
    # via streamlit-extras
prometheus-client==0.21.1
    # via streamlit-extras
protobuf==5.29.4
    # via
    #   streamlit
    #   streamlit-extras
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
pyarrow==19.0.1
    # via streamlit
pydeck==0.9.1
    # via streamlit
pymdown-extensions==10.14.3
    # via markdownlit
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jsii
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via pymdown-extensions
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   favicon
    #   streamlit
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.11.4
    # via boto3
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
soupsieve==2.6
    # via beautifulsoup4
st-annotated-text==4.0.2
    # via streamlit-extras
st-theme==1.2.3
    # via streamlit-extras
streamlit==1.44.1
    # via
    #   bedrock-knowledge-bot (pyproject.toml)
    #   markdownlit
    #   st-theme
    #   streamlit-avatar
    #   streamlit-camera-input-live
    #   streamlit-card
    #   streamlit-embedcode
    #   streamlit-extras
    #   streamlit-faker
    #   streamlit-image-coordinates
    #   streamlit-keyup
    #   streamlit-toggle-switch
    #   streamlit-vertical-slider
streamlit-avatar==0.1.3
    # via streamlit-extras
streamlit-camera-input-live==0.2.0
    # via streamlit-extras
streamlit-card==1.0.2
    # via streamlit-extras
streamlit-embedcode==0.1.2
    # via streamlit-extras
streamlit-extras==0.6.0
    # via
    #   bedrock-knowledge-bot (pyproject.toml)
    #   markdownlit
    #   streamlit-faker
streamlit-faker==0.0.3
    # via streamlit-extras
streamlit-image-coordinates==0.1.9
    # via streamlit-extras
streamlit-keyup==0.3.0
    # via streamlit-extras
streamlit-toggle-switch==1.0.2
    # via streamlit-extras
streamlit-vertical-slider==2.5.5
    # via streamlit-extras
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.13.2
    # via
    #   altair
    #   beautifulsoup4
    #   jsii
    #   streamlit
tzdata==2025.2
    # via
    #   faker
    #   pandas
urllib3==2.4.0
    # via
    #   botocore
    #   requests
validators==0.34.0
    # via streamlit-extras
